+++ examples/minimize_curve_length
@@ -2,7 +2,7 @@
 
 import numpy as np
 from simsopt.geo.curverzfourier import CurveRZFourier
-from simsopt.geo.objectives import CurveLength
+from simsopt.geo.curveobjectives import CurveLength
 from simsopt.core.optimizable import optimizable
 from simsopt.core.least_squares_problem import LeastSquaresProblem
 from simsopt.solve.serial_solve import least_squares_serial_solve
