+++ src/simsopt/geo/curveobjectives.py
